Version 1.0.1013

+++ GLChip.m
@@ -188,7 +188,7 @@
     stackVectors[offsetVector - 2].z -= 3 * (1 - fade);
     stackVectors[offsetVector - 1].z -= 3 * (1 - fade);
     
-    NSLog(@"%X - %X - %f5", self, self.chipGroup, self.chipGroup.opacity);
+    NSLog(@"%X - %X - %f5", (uint)self, (uint)self.chipGroup, self.chipGroup.opacity);
     
     stackColors[offsetColors - 8] = Color3DMake(lightness * fade * self.chipGroup.opacity, lightness * fade * self.chipGroup.opacity, lightness * fade * self.chipGroup.opacity,             fade * self.chipGroup.opacity);
     stackColors[offsetColors - 7] = Color3DMake(lightness * fade * self.chipGroup.opacity, lightness * fade * self.chipGroup.opacity, lightness * fade * self.chipGroup.opacity,             fade * self.chipGroup.opacity);
